requirements.txt
# Core PDF and Image Processing
PyMuPDF==1.23.8
opencv-python==4.8.1.78
Pillow==10.1.0
pdf2image==1.16.3

# OCR Libraries
easyocr==1.7.0
pytesseract==0.3.10

# Computer Vision and Image Processing
imagehash==4.3.1
scikit-image==0.22.0

# Database and ORM
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Processing and Analysis
pandas==2.1.4
numpy==1.25.2

# API and Web Requests
requests==2.31.0
urllib3==2.1.0

# String Matching and Text Processing
python-Levenshtein==0.21.1
fuzzywuzzy==0.18.0

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CLI and Utilities
click==8.1.7
tqdm==4.66.1

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0

# Development utilities (optional)
black==23.11.0
flake8==6.1.0